Add support for environment variables.
If a config.ini setting starts with a $, it is replaced with the
environment variable by the name of that value (without the $).

This enables Jig to more easily be run on Heroku, since heroku wants all
of the web workers to listen on a particular port, for load balancing
reasons.